requirements.txt
# tensorflow
#tensorflow-cpu==2.10.0
herbie-data
#numpy==1.21.6
numpy
pandas
requests
scikit-learn
#xarray<2023.12
xarray
matplotlib
cfgrib
cartopy
pytz
astral